deploy to local maven repo
It turns out that sbt doesn't handle well the situation in which the
recursive dependencies include two different versions of the same github
project. That's probably because github versions are indicated using
opaque tags and branch names, not via semantically-significant version
numbers which sbt could compare in order to figure out which one is the
most recent version.

So we now publish scala-utils to our local maven repo, so it is
published with a semantic version number which sbt can understand. The
downside is that since our local maven repo is not accessible from
outside KeaText, open-source contributor who wants to contribute to
scala-utils wouldn't be able to run "sbt publish". I don't think that's
a problem, because when you're an ordinary contributor to an open source
repo like akka, you don't usually have the permission to publish a new
version of akka either, you rely on official maintainers to do that.